extract shared layer UT fixtures into layer_ut_utils
